Move footer into layout folder
