+++ expense-splitter/src/App.jsx
@@ -1,6 +1,6 @@
 import Expense from './components/expense/Expense'
 import Group from './components/group/Group'
-import Footer from './components/Footer'
+import Footer from './components/layout/Footer'
 import { DataProvider } from './components/context/SiteContext'
 import Header from './components/layout/Header'
