+++ expense-splitter/src/components/layout/Footer.jsx
